move ElevatorState to elevatorSubsystem package
