+++ src/main/scheduler/ElevatorMonitor.java
@@ -5,6 +5,7 @@ import java.util.HashSet;
 import java.util.Iterator;
 import java.util.LinkedHashSet;
 
+import main.elevatorSubsystem.ElevatorState;
 import main.global.Direction;
 import main.global.ElevatorDoorStatus;
 import main.global.ElevatorStatus;
