+++ src/main/scheduler/Scheduler.java
@@ -8,7 +8,6 @@ import java.util.HashSet;
 import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.Queue;
-import java.util.TreeSet;
 
 import main.ElevatorSystemComponent;
 import main.global.Direction;
